The great and powerful ABI-breaker; changing many, many ints to 'narrower' in use variables: - boolean things into 1-bit fields, - magical bools to 2 or more (and dev w/ first look sees 	that field has some magic), - timedate things into time_t, - sizes, counts into size_t, - offsets, indexes into off_t (not all, exactly), - some enum-defined values into correct enums.
Also broke magical 'xstate' field into 2 one-bit fields.
XXX: think about magical XSTATE query.


What do you think about switching boolean args to functions
into bool? Basic motivation: to make dev be sure, whether
that arg is 0/1, or it is another thing of magical whatever.

+++ ekg/strings.c
@@ -43,11 +43,11 @@
 /* stringo-naprawiacz, taki jak ufix() w xmalloc */
 #define ufix(x)	((wchar_t *) x ? (wchar_t *) x : (wchar_t *) L"")
 
-inline int xwcslen(const CHAR_T *str) {
+inline size_t xwcslen(const CHAR_T *str) {
 	return wcslen(ufix(str));
 }
 
-inline int xmbslen(const char *str) {
+inline size_t xmbslen(const char *str) {
 	return mbstowcs(NULL, str ? str : "", 0);
 }
